refact: abstract ExpensePayeeDetails component
